Code reorg - PiwikPDFExporter uses the same source
as PiwikStatisticsReader.
In addition it can now pull the handle-country data from the cache
server.
