+++ dspace-api/src/main/java/cz/cuni/mff/ufal/PiwikHelper.java
@@ -1,117 +1,39 @@
 package cz.cuni.mff.ufal;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.StringReader;
-import java.io.StringWriter;
-import java.io.Writer;
+import java.io.*;
 import java.net.URL;
-import java.util.Set;
+import java.net.URLEncoder;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.*;
 
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
-import javax.xml.transform.OutputKeys;
-import javax.xml.transform.Transformer;
-import javax.xml.transform.TransformerFactory;
-import javax.xml.transform.dom.DOMSource;
-import javax.xml.transform.stream.StreamResult;
+import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.xpath.*;
 
+import org.apache.tools.ant.filters.StringInputStream;
+import org.dspace.content.Item;
+import org.dspace.core.ConfigurationManager;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
-import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
-import org.xml.sax.InputSource;
 
 public class PiwikHelper {
 	private static org.apache.log4j.Logger log = Logger.getLogger(PiwikHelper.class);
 
-	private static final String[] requiredStatsNames = {"nb_visits", "nb_uniq_visitors", "nb_pageviews",
-			"nb_uniq_pageviews", "nb_downloads", "nb_uniq_downloads"};
+	private String dspaceURL = ConfigurationManager.getProperty("dspace.url");
+	/** Piwik configurations */
+	private static final String PIWIK_API_MODE = ConfigurationManager.getProperty("lr", "lr.statistics.api.mode");
+	private static final String PIWIK_API_URL = ConfigurationManager.getProperty("lr", "lr.statistics.api.url");
+	private static final String PIWIK_API_URL_CACHED = ConfigurationManager.getProperty("lr", "lr.statistics.api.cached.url");
+	private static final String PIWIK_AUTH_TOKEN = ConfigurationManager.getProperty("lr", "lr.statistics.api.auth.token");
+	private static final String PIWIK_SITE_ID = ConfigurationManager.getProperty("lr", "lr.statistics.api.site_id");
+	private static final String PIWIK_DOWNLOAD_SITE_ID = ConfigurationManager.getProperty("lr", "lr.tracker.bitstream.site_id");
 
-
-	public static String mergeXML(String report, String downloadReport) throws Exception {
-		/**
-		 * add page views from downloadReport as nb_downloads to report
-		 */
-		Document reportDoc = loadXMLFromString(report);
-		Document downloadReportDoc = loadXMLFromString(downloadReport);
-		
-		XPathFactory xpathFactory = XPathFactory.newInstance();
-		XPath xpath = xpathFactory.newXPath();
-		XPathExpression resExpr = xpath.compile("//result");
-		XPathExpression downExpr = xpath.compile("./nb_downloads");
-		XPathExpression uniqDownExpr = xpath.compile("./nb_uniq_downloads");
-				
-		NodeList rRows = (NodeList)resExpr.evaluate(reportDoc, XPathConstants.NODESET);
-
-		for(int i=0;i<rRows.getLength();i++) {
-			Node rRow = rRows.item(i);
-			rRow = ensureRequiredStatsPresent(reportDoc, xpath, rRow);
-			Node down = (Node) downExpr.evaluate(rRow, XPathConstants.NODE);
-			Node uniqDown = (Node) uniqDownExpr.evaluate(rRow, XPathConstants.NODE);
-			XPathExpression pvExpr = xpath.compile(String.format("/results/result[@%s]/nb_pageviews/text()", rRow.getAttributes().getNamedItem("date")));
-			XPathExpression upvExpr = xpath.compile(String.format("/results/result[@%s]/nb_uniq_pageviews/text()", rRow.getAttributes().getNamedItem("date")));
-
-			int pvCount = ((Double)pvExpr.evaluate(downloadReportDoc, XPathConstants.NUMBER)).intValue();
-			int upvCount = ((Double)upvExpr.evaluate(downloadReportDoc, XPathConstants.NUMBER)).intValue();
-
-			down.setTextContent(Integer.toString(pvCount));
-			uniqDown.setTextContent(Integer.toString(upvCount));
-		}
-
-		Transformer tf = TransformerFactory.newInstance().newTransformer();
-		tf.setOutputProperty(OutputKeys.ENCODING, "UTF-8");
-		tf.setOutputProperty(OutputKeys.INDENT, "yes");
-		Writer out = new StringWriter();
-		tf.transform(new DOMSource(reportDoc), new StreamResult(out));
-		return out.toString();
-	}
-
-	private static Node ensureRequiredStatsPresent(Document reportDoc, XPath xpath, Node rRow) throws XPathExpressionException {
-		for(String tagName : requiredStatsNames){
-			XPathExpression expr = xpath.compile("./" + tagName);
-			Node node = (Node) expr.evaluate(rRow, XPathConstants.NODE);
-			if(node == null){
-				Element el = reportDoc.createElement(tagName);
-				el.setTextContent("0");
-				rRow.appendChild(el);
-			}
-		}
-		return rRow;
-	}
-
-	public static String mergeJSON(String report, String downloadReport) throws Exception {
-		/**
-		 * add page views from downloadReport as nb_downloads to report
-		 */
-		JSONParser parser = new JSONParser();
-		JSONObject reportJSON = (JSONObject)parser.parse(report);
-		JSONObject downloadReportJSON = (JSONObject)parser.parse(downloadReport);
-		for(Object key : reportJSON.keySet()) {			
-			JSONObject rRow = null;
-			JSONObject dRow = null;
-			try{
-				dRow = (JSONObject)downloadReportJSON.get(key);
-			} catch (ClassCastException e) {
-				continue;
-			}
-			try {
-				rRow = (JSONObject)reportJSON.get(key);
-			} catch (ClassCastException e) {
-				rRow = new JSONObject();
-				reportJSON.put(key, rRow);
-			}			
-			rRow.put("nb_downloads", dRow.get("nb_pageviews"));
-			rRow.put("nb_uniq_downloads", dRow.get("nb_uniq_pageviews"));
-		}
-		return reportJSON.toJSONString();
-	}
-	
 	/**
 	 *
 	 * @param keys - The order of keys should match the json object indexes
@@ -434,11 +356,272 @@ public class PiwikHelper {
 		return output.toString();
 	}
 
-	public static Document loadXMLFromString(String xml) throws Exception {
-	    DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
-	    DocumentBuilder builder = factory.newDocumentBuilder();
-	    InputSource is = new InputSource(new StringReader(xml));
-	    return builder.parse(is);
+	String period;
+	String date;
+	Item item;
+	String rest;
+	public PiwikHelper(String period, String date, Item item, String rest){
+		this.period = period;
+		this.date = date;
+		this.item = item;
+		this.rest = rest;
 	}
-		
+
+	public String getDataAsJsonString() throws Exception {
+		String mergedResult;
+		if(PIWIK_API_MODE.equals("cached")) {
+			log.debug("========CACHED MODE");
+			mergedResult = getDataFromLindatPiwikCacheServer();
+		} else {
+			// direct mode as default
+			log.debug("========DIRECT MODE");
+			mergedResult = getDataFromPiwikServer();
+		}
+		return mergedResult;
+	}
+
+	public List<String[]> getCountryData() throws Exception{
+		if(PIWIK_API_MODE.equals("cached")) {
+			log.debug("========CACHED MODE");
+			return getCountryDataFromLindatPiwikCacheServer();
+		} else {
+			// direct mode as default
+			log.debug("========DIRECT MODE");
+			return getCountryDataFromPiwik();
+		}
+	}
+
+	private List<String[]> getCountryDataFromLindatPiwikCacheServer() throws Exception {
+		String url = PIWIK_API_URL_CACHED + "handle?h=" + item.getHandle() + "&period=month&country=true";
+
+		if(date!=null) {
+			url += "&date=" + date;
+		}
+		JSONParser parser = new JSONParser();
+		JSONObject countriesReport = (JSONObject)parser.parse(PiwikHelper.readFromURL(url));
+
+		List<String[]> result = new ArrayList<>(10);
+
+		for (Object key : countriesReport.keySet() ) {
+			if(key instanceof String){
+				String date = (String) key;
+				JSONArray countryData = (JSONArray) countriesReport.get(date);
+				for(Object country: countryData){
+					if(country instanceof JSONObject){
+						JSONObject c = (JSONObject) country;
+						String label = (String) c.get("label");
+						String count = ((Integer) c.get("nb_visits")).toString();
+						result.add(new String[]{label, count});
+					}
+				}
+				// expecting only one date key
+				break;
+			}
+		}
+		return result;
+	}
+
+	private List<String[]> getCountryDataFromPiwik() throws Exception {
+
+		String countryReportURL = PIWIK_API_URL + "index.php"
+				+ "?module=API"
+				+ "&method=UserCountry.getCountry"
+				+ "&idSite=" + PIWIK_SITE_ID
+				+ "&period=month"
+				+ "&date=" + date
+				+ "&expanded=1"
+				+ "&token_auth=" + PIWIK_AUTH_TOKEN
+				+ "&filter_limit=10"
+				+ "&format=xml"
+				+ "&segment=pageUrl=@" + item.getHandle();
+
+
+		String xml = PiwikHelper.readFromURL(countryReportURL);
+
+		Document doc = parseXML(xml);
+
+		if(doc==null) throw new Exception("Unable to parse XML");
+
+		List<String[]> data = new ArrayList<String[]>();
+
+		XPath xPath =  XPathFactory.newInstance().newXPath();
+		XPathExpression eachResultNode = xPath.compile("//result/row");
+
+		NodeList results = (NodeList)eachResultNode.evaluate(doc, XPathConstants.NODESET);
+
+		for(int i=0;i<results.getLength();i++) {
+			Element row = (Element)results.item(i);
+			String country = row.getElementsByTagName("label").item(0).getTextContent();
+			String count = row.getElementsByTagName("nb_visits").item(0).getTextContent();
+			data.add(new String[]{country, count});
+		}
+
+		return data;
+
+	}
+
+	private static Document parseXML(String xml) {
+		DocumentBuilderFactory builderFactory = DocumentBuilderFactory.newInstance();
+		DocumentBuilder builder = null;
+		Document doc = null;
+		try {
+			builder = builderFactory.newDocumentBuilder();
+		} catch (ParserConfigurationException e) {
+			e.printStackTrace();
+		}
+		try {
+			doc = builder.parse(new StringInputStream(xml));
+		} catch (Exception e) {
+			log.error(e);
+		}
+		return doc;
+	}
+
+
+	private String getDataFromLindatPiwikCacheServer() throws IOException {
+		String url = PIWIK_API_URL_CACHED + "handle?h=" + item.getHandle() + "&period=" + period;
+
+		if(date!=null) {
+			url += "&date=" + date;
+		}
+
+		return PiwikHelper.readFromURL(url);
+	}
+
+	private String getDataFromPiwikServer() throws Exception {
+		SortedMap<String, String> urls = buildViewsURL();
+		urls.put("downloads", buildDownloadsURL());
+		String bulkApiGetRequestURL = buildBulkApiGetRequestURL(urls);
+
+		log.debug(String.format("Fetching data from piwik server; requesting \"%s\"", bulkApiGetRequestURL));
+
+		String report = PiwikHelper.readFromURL(bulkApiGetRequestURL);
+		return PiwikHelper.transformJSONResults(urls.keySet(), report);
+	}
+	private String buildBulkApiGetRequestURL(SortedMap<String, String> urls){
+		String piwikBulkApiGetQuery = "module=API&method=API.getBulkRequest&format=JSON"
+				+ "&token_auth=" + PIWIK_AUTH_TOKEN;
+		StringBuilder sb = new StringBuilder();
+		sb.append(PIWIK_API_URL)
+				.append(rest)
+				.append("?")
+				.append(piwikBulkApiGetQuery);
+		int i = 0;
+		for(String url : urls.values()){
+			sb.append("&urls[")
+					.append(i++)
+					.append("]=")
+					.append(url);
+		}
+		return sb.toString();
+	}
+
+	private SortedMap<String, String> buildViewsURL() throws UnsupportedEncodingException, ParseException {
+		// use Actions.getPageUrl; call it twice; once with ?show=full
+		String paramsFmt = "method=Actions.getPageUrl&pageUrl=%s";
+		String summaryItemView = buildURL(PIWIK_SITE_ID,
+				String.format(paramsFmt, URLEncoder.encode(dspaceURL + "/handle/" + item.getHandle(), "UTF-8")));
+		String fullItemView = buildURL(PIWIK_SITE_ID,
+				String.format(paramsFmt, URLEncoder.encode(
+						dspaceURL + "/handle" + "/" + item.getHandle() + "?show=full", "UTF-8")));
+		SortedMap<String, String> ret = new TreeMap<>();
+		ret.put("summaryItemView", summaryItemView);
+		ret.put("fullItemView", fullItemView);
+		return ret;
+	}
+
+	private String buildDownloadsURL() throws UnsupportedEncodingException, ParseException {
+		String filterPattern =  URLEncoder.encode(dspaceURL + "/bitstream/handle/" + item.getHandle(), "UTF-8");
+		String params =
+				"method=Actions.getPageUrls" +
+						"&expanded=1&flat=1" +
+						"&filter_column=url" +
+						"&filter_pattern=" + filterPattern;
+		return buildURL(PIWIK_DOWNLOAD_SITE_ID, params);
+	}
+
+	private String buildURL(String siteID, String specificParams) throws UnsupportedEncodingException,
+			ParseException {
+		String dateRange = DateRange.fromDateString(date).toString();
+
+		/*
+		The Actions API lets you request reports for all your Visitor Actions: Page URLs, Page titles, Events,
+		Content Tracking, File Downloads and Clicks on external websites.
+		Actions.getPageUrls:
+		 - stats(nb_visits, nb_hits, etc) per url, in given date broken down by period
+		expanded:
+		- some API functions have a parameter 'expanded'. If 'expanded' is set to 1, the returned data will contain the first level results, as well as all sub-tables.
+		- basically fetches subtable (if present as idsubdatatable)
+		- eg. urls broken down by directory structure:
+			- lvl 1: <segment>pageUrl=^https%253A%252F%252Fdivezone.net%252Fdiving</segment>
+			- lvl 2: <segment>pageUrl==https%253A%252F%252Fdivezone.net%252Fdiving%252Fbali</segment>; <segment>pageUrl==https%253A%252F%252Fdivezone.net%252Fdiving%252Fthailand</segment>
+			- lvl 1 has no url, ie. it's cummulative for the "underlying" urls
+		flat:
+		- some API functions have a parameter 'expanded', which means that the data is hierarchical. For such API function, if 'flat' is set to 1, the returned data will contain the flattened view of the table data set. The children of all first level rows will be aggregated under one row. This is useful for example to see all Custom Variables names and values at once, for example, Matomo forum user status, or to see the full URLs not broken down by directory or structure.
+		- this will remove the cummulative results; all rows will be of the same type (having url field)
+		 */
+		String params =
+				specificParams
+						+ "&date=" + dateRange
+						+ "&period=" + period
+						+ "&token_auth=" + PIWIK_AUTH_TOKEN
+						+ "&showColumns=label,url,nb_visits,nb_hits"
+						// don't want to handle "paging" (summary views)
+						+ "&filter_limit=-1"
+						+ "&idSite=" + siteID;
+		return URLEncoder.encode(params, "UTF-8");
+	}
+
+	private static class DateRange {
+		private static final SimpleDateFormat df = new SimpleDateFormat("yyyy-MM-dd");
+
+		private final Date startDate;
+		private final Date endDate;
+
+		private DateRange(Date startDate, Date endDate){
+			this.startDate = startDate;
+			this.endDate = endDate;
+		}
+
+		private static DateRange fromDateString(String date) throws ParseException {
+			Date startDate;
+			Date endDate;
+			if(date!=null) {
+				String sdate = date;
+				String edate = date;
+				if(sdate.length()==4) {
+					sdate += "-01-01";
+					edate += "-12-31";
+				} else if(sdate.length()==7) {
+					Calendar cal = Calendar.getInstance();
+					cal.set(Calendar.YEAR, Integer.parseInt(sdate.substring(0,4)));
+					cal.set(Calendar.MONTH, Integer.parseInt(sdate.substring(5,7))-1);
+					cal.set(Calendar.DATE, cal.getActualMaximum(Calendar.DATE));
+					sdate += "-01";
+					edate = df.format(cal.getTime());
+				}
+				startDate = df.parse(sdate);
+				endDate = df.parse(edate);
+			} else {
+				// default start and end data
+				startDate = df.parse("2014-01-01");
+				endDate = Calendar.getInstance().getTime();
+			}
+			return new DateRange(startDate, endDate);
+		}
+
+		public String getFormattedStartDate(){
+			return df.format(startDate);
+		}
+
+		public String getFormattedEndDate(){
+			return df.format(endDate);
+		}
+
+		@Override
+		public String toString(){
+			return getFormattedStartDate() + "," + getFormattedEndDate();
+		}
+
+	}
 }
