--- dspace-api/src/test/resources/piwik/download_report.json
@@ -1,1 +1,0 @@
-{"2015-07-27":{"nb_uniq_visitors":2,"nb_visits":2,"nb_users":0,"nb_actions":5,"max_actions":3,"bounce_count":0,"sum_visit_length":197,"nb_conversions":0,"nb_visits_converted":0,"revenue":0,"conversion_rate":"0%","nb_conversions_new_visit":0,"nb_visits_converted_new_visit":0,"revenue_new_visit":0,"conversion_rate_new_visit":"0%","nb_pageviews":5,"nb_uniq_pageviews":3,"nb_downloads":0,"nb_uniq_downloads":0,"nb_outlinks":0,"nb_uniq_outlinks":0,"nb_searches":0,"nb_keywords":0,"bounce_rate":"0%","nb_actions_per_visit":2.5,"avg_time_on_site":99,"avg_time_on_site_returning":0,"nb_actions_per_visit_returning":0,"bounce_rate_returning":"0%"},"2015-07-28":{"nb_uniq_visitors":10,"nb_visits":10,"nb_users":0,"nb_actions":17,"max_actions":5,"bounce_count":8,"sum_visit_length":63,"nb_conversions":0,"nb_visits_converted":0,"revenue":0,"conversion_rate":"0%","nb_conversions_new_visit":0,"nb_visits_converted_new_visit":0,"revenue_new_visit":0,"conversion_rate_new_visit":"0%","nb_pageviews":17,"nb_uniq_pageviews":13,"nb_downloads":0,"nb_uniq_downloads":0,"nb_outlinks":0,"nb_uniq_outlinks":0,"nb_searches":0,"nb_keywords":0,"bounce_rate":"80%","nb_actions_per_visit":1.7,"avg_time_on_site":6,"avg_time_on_site_returning":0,"nb_actions_per_visit_returning":0,"bounce_rate_returning":"0%"},"2015-07-29":{"nb_uniq_visitors":4,"nb_visits":4,"nb_users":0,"nb_actions":4,"max_actions":1,"bounce_count":4,"sum_visit_length":0,"nb_conversions":0,"nb_visits_converted":0,"revenue":0,"conversion_rate":"0%","nb_conversions_new_visit":0,"nb_visits_converted_new_visit":0,"revenue_new_visit":0,"conversion_rate_new_visit":"0%","nb_pageviews":4,"nb_uniq_pageviews":4,"nb_downloads":0,"nb_uniq_downloads":0,"nb_outlinks":0,"nb_uniq_outlinks":0,"nb_searches":0,"nb_keywords":0,"bounce_rate":"100%","nb_actions_per_visit":1,"avg_time_on_site":0,"avg_time_on_site_returning":0,"nb_actions_per_visit_returning":0,"bounce_rate_returning":"0%"},"2015-07-30":{"nb_uniq_visitors":5,"nb_visits":5,"nb_users":0,"nb_actions":14,"max_actions":4,"bounce_count":1,"sum_visit_length":284,"nb_conversions":0,"nb_visits_converted":0,"revenue":0,"conversion_rate":"0%","nb_conversions_new_visit":0,"nb_visits_converted_new_visit":0,"revenue_new_visit":0,"conversion_rate_new_visit":"0%","nb_pageviews":14,"nb_uniq_pageviews":10,"nb_downloads":0,"nb_uniq_downloads":0,"nb_outlinks":0,"nb_uniq_outlinks":0,"nb_searches":0,"nb_keywords":0,"bounce_rate":"20%","nb_actions_per_visit":2.8,"avg_time_on_site":57,"avg_time_on_site_returning":0,"nb_actions_per_visit_returning":0,"bounce_rate_returning":"0%"},"2015-07-31":{"nb_uniq_visitors":4,"nb_visits":4,"nb_users":0,"nb_actions":4,"max_actions":1,"bounce_count":4,"sum_visit_length":0,"nb_conversions":0,"nb_visits_converted":0,"revenue":0,"conversion_rate":"0%","nb_conversions_new_visit":0,"nb_visits_converted_new_visit":0,"revenue_new_visit":0,"conversion_rate_new_visit":"0%","nb_pageviews":4,"nb_uniq_pageviews":4,"nb_downloads":0,"nb_uniq_downloads":0,"nb_outlinks":0,"nb_uniq_outlinks":0,"nb_searches":0,"nb_keywords":0,"bounce_rate":"100%","nb_actions_per_visit":1,"avg_time_on_site":0,"avg_time_on_site_returning":0,"nb_actions_per_visit_returning":0,"bounce_rate_returning":"0%"},"2015-08-01":[],"2015-08-02":{"nb_uniq_visitors":1,"nb_visits":1,"nb_users":0,"nb_actions":1,"max_actions":1,"bounce_count":1,"sum_visit_length":0,"nb_conversions":0,"nb_visits_converted":0,"revenue":0,"conversion_rate":"0%","nb_conversions_new_visit":0,"nb_visits_converted_new_visit":0,"revenue_new_visit":0,"conversion_rate_new_visit":"0%","nb_pageviews":1,"nb_uniq_pageviews":1,"nb_downloads":0,"nb_uniq_downloads":0,"nb_outlinks":0,"nb_uniq_outlinks":0,"nb_searches":0,"nb_keywords":0,"bounce_rate":"100%","nb_actions_per_visit":1,"avg_time_on_site":0,"avg_time_on_site_returning":0,"nb_actions_per_visit_returning":0,"bounce_rate_returning":"0%"},"2015-08-03":{"nb_uniq_visitors":4,"nb_visits":4,"nb_users":0,"nb_actions":6,"max_actions":3,"bounce_count":3,"sum_visit_length":218,"nb_conversions":0,"nb_visits_converted":0,"revenue":0,"conversion_rate":"0%","nb_conversions_new_visit":0,"nb_visits_converted_new_visit":0,"revenue_new_visit":0,"conversion_rate_new_visit":"0%","nb_pageviews":6,"nb_uniq_pageviews":6,"nb_downloads":0,"nb_uniq_downloads":0,"nb_outlinks":0,"nb_uniq_outlinks":0,"nb_searches":0,"nb_keywords":0,"bounce_rate":"75%","nb_actions_per_visit":1.5,"avg_time_on_site":55,"avg_time_on_site_returning":0,"nb_actions_per_visit_returning":0,"bounce_rate_returning":"0%"}}
