--- dspace-api/src/test/resources/piwik/view_report.json
@@ -1,1 +1,0 @@
-{"2015-07-27":{"nb_uniq_visitors":3,"nb_visits":3,"nb_users":0,"nb_actions":8,"max_actions":4,"bounce_count":1,"sum_visit_length":545,"nb_conversions":0,"nb_visits_converted":0,"revenue":0,"conversion_rate":"0%","nb_conversions_new_visit":0,"nb_visits_converted_new_visit":0,"revenue_new_visit":0,"conversion_rate_new_visit":"0%","nb_pageviews":3,"nb_uniq_pageviews":3,"nb_downloads":0,"nb_uniq_downloads":0,"nb_outlinks":0,"nb_uniq_outlinks":0,"nb_searches":0,"nb_keywords":0,"nb_hits_with_time_generation":3,"avg_time_generation":1.16,"bounce_rate":"33%","nb_actions_per_visit":2.7,"avg_time_on_site":182,"avg_time_on_site_returning":0,"nb_actions_per_visit_returning":0,"bounce_rate_returning":"0%"},"2015-07-28":{"nb_uniq_visitors":7,"nb_visits":9,"nb_users":0,"nb_actions":43,"max_actions":24,"bounce_count":4,"sum_visit_length":1949,"nb_visits_returning":5,"nb_actions_returning":13,"nb_uniq_visitors_returning":3,"nb_users_returning":0,"max_actions_returning":8,"bounce_rate_returning":"60%","nb_actions_per_visit_returning":2.6,"avg_time_on_site_returning":142,"nb_conversions":0,"nb_visits_converted":0,"revenue":0,"conversion_rate":"0%","nb_conversions_new_visit":0,"nb_visits_converted_new_visit":0,"revenue_new_visit":0,"conversion_rate_new_visit":"0%","nb_conversions_returning_visit":0,"nb_visits_converted_returning_visit":0,"revenue_returning_visit":0,"conversion_rate_returning_visit":"0%","nb_pageviews":10,"nb_uniq_pageviews":9,"nb_downloads":0,"nb_uniq_downloads":0,"nb_outlinks":0,"nb_uniq_outlinks":0,"nb_searches":0,"nb_keywords":0,"nb_hits_with_time_generation":10,"avg_time_generation":0.443,"bounce_rate":"44%","nb_actions_per_visit":4.8,"avg_time_on_site":217},"2015-07-29":{"nb_uniq_visitors":6,"nb_visits":6,"nb_users":0,"nb_actions":10,"max_actions":3,"bounce_count":3,"sum_visit_length":63,"nb_visits_returning":2,"nb_actions_returning":3,"nb_uniq_visitors_returning":2,"nb_users_returning":0,"max_actions_returning":2,"bounce_rate_returning":"50%","nb_actions_per_visit_returning":1.5,"avg_time_on_site_returning":5,"nb_conversions":0,"nb_visits_converted":0,"revenue":0,"conversion_rate":"0%","nb_conversions_new_visit":0,"nb_visits_converted_new_visit":0,"revenue_new_visit":0,"conversion_rate_new_visit":"0%","nb_conversions_returning_visit":0,"nb_visits_converted_returning_visit":0,"revenue_returning_visit":0,"conversion_rate_returning_visit":"0%","nb_pageviews":7,"nb_uniq_pageviews":6,"nb_downloads":0,"nb_uniq_downloads":0,"nb_outlinks":0,"nb_uniq_outlinks":0,"nb_searches":0,"nb_keywords":0,"nb_hits_with_time_generation":5,"avg_time_generation":0.416,"bounce_rate":"50%","nb_actions_per_visit":1.7,"avg_time_on_site":11},"2015-07-30":{"nb_uniq_visitors":7,"nb_visits":10,"nb_users":0,"nb_actions":55,"max_actions":28,"bounce_count":3,"sum_visit_length":3324,"nb_visits_returning":5,"nb_actions_returning":16,"nb_uniq_visitors_returning":3,"nb_users_returning":0,"max_actions_returning":4,"bounce_rate_returning":"20%","nb_actions_per_visit_returning":3.2,"avg_time_on_site_returning":299,"nb_conversions":0,"nb_visits_converted":0,"revenue":0,"conversion_rate":"0%","nb_conversions_new_visit":0,"nb_visits_converted_new_visit":0,"revenue_new_visit":0,"conversion_rate_new_visit":"0%","nb_conversions_returning_visit":0,"nb_visits_converted_returning_visit":0,"revenue_returning_visit":0,"conversion_rate_returning_visit":"0%","nb_pageviews":17,"nb_uniq_pageviews":10,"nb_downloads":0,"nb_uniq_downloads":0,"nb_outlinks":0,"nb_uniq_outlinks":0,"nb_searches":0,"nb_keywords":0,"nb_hits_with_time_generation":17,"avg_time_generation":0.708,"bounce_rate":"30%","nb_actions_per_visit":5.5,"avg_time_on_site":332},"2015-07-31":{"nb_uniq_visitors":4,"nb_visits":4,"nb_users":0,"nb_actions":6,"max_actions":2,"bounce_count":2,"sum_visit_length":25,"nb_visits_returning":1,"nb_actions_returning":2,"nb_uniq_visitors_returning":1,"nb_users_returning":0,"max_actions_returning":2,"bounce_rate_returning":"0%","nb_actions_per_visit_returning":2,"avg_time_on_site_returning":7,"nb_conversions":0,"nb_visits_converted":0,"revenue":0,"conversion_rate":"0%","nb_conversions_new_visit":0,"nb_visits_converted_new_visit":0,"revenue_new_visit":0,"conversion_rate_new_visit":"0%","nb_conversions_returning_visit":0,"nb_visits_converted_returning_visit":0,"revenue_returning_visit":0,"conversion_rate_returning_visit":"0%","nb_pageviews":4,"nb_uniq_pageviews":4,"nb_downloads":0,"nb_uniq_downloads":0,"nb_outlinks":0,"nb_uniq_outlinks":0,"nb_searches":0,"nb_keywords":0,"nb_hits_with_time_generation":4,"avg_time_generation":1.323,"bounce_rate":"50%","nb_actions_per_visit":1.5,"avg_time_on_site":6},"2015-08-01":[],"2015-08-02":{"nb_uniq_visitors":2,"nb_visits":2,"nb_users":0,"nb_actions":6,"max_actions":5,"bounce_count":1,"sum_visit_length":63,"nb_visits_returning":1,"nb_actions_returning":5,"nb_uniq_visitors_returning":1,"nb_users_returning":0,"max_actions_returning":5,"bounce_rate_returning":"0%","nb_actions_per_visit_returning":5,"avg_time_on_site_returning":63,"nb_conversions":0,"nb_visits_converted":0,"revenue":0,"conversion_rate":"0%","nb_conversions_new_visit":0,"nb_visits_converted_new_visit":0,"revenue_new_visit":0,"conversion_rate_new_visit":"0%","nb_conversions_returning_visit":0,"nb_visits_converted_returning_visit":0,"revenue_returning_visit":0,"conversion_rate_returning_visit":"0%","nb_pageviews":2,"nb_uniq_pageviews":2,"nb_downloads":0,"nb_uniq_downloads":0,"nb_outlinks":0,"nb_uniq_outlinks":0,"nb_searches":0,"nb_keywords":0,"nb_hits_with_time_generation":2,"avg_time_generation":0.375,"bounce_rate":"50%","nb_actions_per_visit":3,"avg_time_on_site":32},"2015-08-03":{"nb_uniq_visitors":8,"nb_visits":8,"nb_users":0,"nb_actions":98,"max_actions":73,"bounce_count":2,"sum_visit_length":5258,"nb_visits_returning":5,"nb_actions_returning":92,"nb_uniq_visitors_returning":5,"nb_users_returning":0,"max_actions_returning":73,"bounce_rate_returning":"20%","nb_actions_per_visit_returning":18.4,"avg_time_on_site_returning":939,"nb_conversions":0,"nb_visits_converted":0,"revenue":0,"conversion_rate":"0%","nb_conversions_new_visit":0,"nb_visits_converted_new_visit":0,"revenue_new_visit":0,"conversion_rate_new_visit":"0%","nb_conversions_returning_visit":0,"nb_visits_converted_returning_visit":0,"revenue_returning_visit":0,"conversion_rate_returning_visit":"0%","nb_pageviews":21,"nb_uniq_pageviews":11,"nb_downloads":0,"nb_uniq_downloads":0,"nb_outlinks":0,"nb_uniq_outlinks":0,"nb_searches":0,"nb_keywords":0,"nb_hits_with_time_generation":18,"avg_time_generation":0.337,"bounce_rate":"25%","nb_actions_per_visit":12.3,"avg_time_on_site":657}}
